requirements.txt
anyio
certifi
charset-normalizer
click
colorama
cycler
fastapi
filelock
fonttools
h11
httpcore
httpx
huggingface-hub
idna
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
packaging
pandas
Pillow
pydantic
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
rfc3986
six
sniffio
safetensors
starlette
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing-extensions
urllib3
uvicorn
wordcloud